requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.4.16
chardet==3.0.4
dj-database-url==0.5.0
Django==2.0.5
future==0.16.0
idna==2.6
parsedatetime==2.4
Pillow==5.1.0
python-decouple==3.1
pytz==2018.4
requests==2.18.4
Unipath==1.1
urllib3==1.22
